+++ packages/client/src/v2-events/features/events/actions/declare/Pages.interaction.stories.tsx
@@ -460,9 +460,7 @@ export const CanSubmitValidlyFilledForm: Story = {
             return undeclaredDraftEvent
           }),
           tRPCMsw.event.actions.declare.request.mutation(async (payload) => {
-            await expect(payload.declaration).not.toHaveProperty(
-              'recommender.name'
-            )
+            await expect(payload.declaration?.['recommender.name']).toBeNull()
             return generateEventDocument({
               configuration: tennisClubMembershipEvent,
               actions: [